backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.8.2
pydantic-settings>=2.3.4
supabase>=2.6.0
pyswisseph>=2.10.3.2
timezonefinder>=6.5.2
pytz>=2024.1
anyio>=4.4.0
openai>=1.35.7
python-dotenv>=1.0.1
httpx>=0.27.0
